Updated the `bytes_to_string` usage.
The function `bytes_to_string` was removed from the recent code and was
replaced by Frame::to_string function, which also has the decoding logic.

This change replaces the bytes_to_string by Frame::to_string calls in
the README.

Additionally, the to_string function always performs the decoding, when
not disabled by the flag. In order to get the raw data in cec-o-matic
format, added a boolean flag to skip the decoding.

The decoding logic from the example was modified to use the built-in
decoding from to_string. To get the original data, pass the skip=true
parameter to the to_string function.